pyproject.toml
[tool.poetry]
name = "monday-app"
version = "0.1.0"
description = ""
authors = ["Daniel Raymond"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
ruff = "^0.5.1"
pyright = "^1.1.370"
python-dotenv = "^1.0.1"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
